@@ -10,8 +10,8 @@
     "antd": "^4.18.9",
     "husky": "^7.0.4",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
@@ -23,8 +23,8 @@
     "@types/antd": "^1.0.0",
     "@types/jest": "^24.9.1",
     "@types/node": "^12.20.46",
-    "@types/react": "^16.14.23",
-    "@types/react-dom": "^16.9.14",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-redux": "^7.1.22",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
     "@typescript-eslint/parser": "^5.13.0",
