@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 
 import { App } from './App/App';
@@ -7,11 +7,11 @@ import { store } from './redux/store';
 
 import './index.css';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root'),
+const container = document.getElementById('root');
+const root = createRoot(container ?? document.body);
+
+root.render(
+  <Provider store={store}>
+    <App />
+  </Provider>,
 );
